# [SPARK-31414][SQL] Fix performance regression with new TimestampFormatter for json and csv time parsing