--- a/sql/core/benchmarks/CSVBenchmark-results.txt
+++ b/sql/core/benchmarks/CSVBenchmark-results.txt
@@ -2,66 +2,66 @@
 Benchmark to measure CSV read/write performance
 ================================================================================================
 
-OpenJDK 64-Bit Server VM 1.8.0_232-8u232-b09-0ubuntu1~18.04.1-b09 on Linux 4.15.0-1044-aws
-Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Parsing quoted values:                    Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-One quoted string                                 51602          51659          59          0.0     1032039.4       1.0X
+One quoted string                                 24073          24109          33          0.0      481463.5       1.0X
 
-OpenJDK 64-Bit Server VM 1.8.0_232-8u232-b09-0ubuntu1~18.04.1-b09 on Linux 4.15.0-1044-aws
-Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Wide rows with 1000 columns:              Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-Select 1000 columns                              191926         192879        1615          0.0      191925.6       1.0X
-Select 100 columns                                46766          46846          69          0.0       46766.1       4.1X
-Select one column                                 35877          35930          83          0.0       35876.8       5.3X
-count()                                           11186          11262          65          0.1       11186.0      17.2X
-Select 100 columns, one bad input field           59943          60107         232          0.0       59943.0       3.2X
-Select 100 columns, corrupt record field          73062          73406         479          0.0       73062.2       2.6X
+Select 1000 columns                               58415          59611        2071          0.0       58414.8       1.0X
+Select 100 columns                                22568          23020         594          0.0       22568.0       2.6X
+Select one column                                 18995          19058          99          0.1       18995.0       3.1X
+count()                                            5301           5332          30          0.2        5300.9      11.0X
+Select 100 columns, one bad input field           39736          40153         361          0.0       39736.1       1.5X
+Select 100 columns, corrupt record field          47195          47826         590          0.0       47195.2       1.2X
 
-OpenJDK 64-Bit Server VM 1.8.0_232-8u232-b09-0ubuntu1~18.04.1-b09 on Linux 4.15.0-1044-aws
-Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Count a dataset with 10 columns:          Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-Select 10 columns + count()                       22389          22447          87          0.4        2238.9       1.0X
-Select 1 column + count()                         14844          14890          43          0.7        1484.4       1.5X
-count()                                            5519           5538          18          1.8         551.9       4.1X
+Select 10 columns + count()                        9884           9904          25          1.0         988.4       1.0X
+Select 1 column + count()                          6794           6835          46          1.5         679.4       1.5X
+count()                                            2060           2065           5          4.9         206.0       4.8X
 
-OpenJDK 64-Bit Server VM 1.8.0_232-8u232-b09-0ubuntu1~18.04.1-b09 on Linux 4.15.0-1044-aws
-Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Write dates and timestamps:               Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-Create a dataset of timestamps                     1949           1977          25          5.1         194.9       1.0X
-to_csv(timestamp)                                 14944          15702         714          0.7        1494.4       0.1X
-write timestamps to files                         12983          12998          14          0.8        1298.3       0.2X
-Create a dataset of dates                          2156           2164           7          4.6         215.6       0.9X
-to_csv(date)                                       9675           9709          41          1.0         967.5       0.2X
-write dates to files                               7880           7897          15          1.3         788.0       0.2X
+Create a dataset of timestamps                      717            732          18         14.0          71.7       1.0X
+to_csv(timestamp)                                  6994           7100         121          1.4         699.4       0.1X
+write timestamps to files                          6417           6435          27          1.6         641.7       0.1X
+Create a dataset of dates                           827            855          24         12.1          82.7       0.9X
+to_csv(date)                                       4408           4438          32          2.3         440.8       0.2X
+write dates to files                               3738           3758          28          2.7         373.8       0.2X
 
-OpenJDK 64-Bit Server VM 1.8.0_232-8u232-b09-0ubuntu1~18.04.1-b09 on Linux 4.15.0-1044-aws
-Intel(R) Xeon(R) CPU E5-2670 v2 @ 2.50GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Read dates and timestamps:                Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-read timestamp text from files                     2235           2245          10          4.5         223.5       1.0X
-read timestamps from files                        54490          54690         283          0.2        5449.0       0.0X
-infer timestamps from files                      104501         104737         236          0.1       10450.1       0.0X
-read date text from files                          2035           2040           6          4.9         203.5       1.1X
-read date from files                              39650          39707          52          0.3        3965.0       0.1X
-infer date from files                             29235          29363         164          0.3        2923.5       0.1X
-timestamp strings                                  3412           3426          18          2.9         341.2       0.7X
-parse timestamps from Dataset[String]             66864          67804         981          0.1        6686.4       0.0X
-infer timestamps from Dataset[String]            118780         119284         837          0.1       11878.0       0.0X
-date strings                                       3730           3734           4          2.7         373.0       0.6X
-parse dates from Dataset[String]                  48728          49071         309          0.2        4872.8       0.0X
-from_csv(timestamp)                               62294          62493         260          0.2        6229.4       0.0X
-from_csv(date)                                    44581          44665         117          0.2        4458.1       0.1X
+read timestamp text from files                     1121           1176          52          8.9         112.1       1.0X
+read timestamps from files                        21298          21366         105          0.5        2129.8       0.1X
+infer timestamps from files                       41008          41051          39          0.2        4100.8       0.0X
+read date text from files                           962            967           5         10.4          96.2       1.2X
+read date from files                              11749          11772          22          0.9        1174.9       0.1X
+infer date from files                             12426          12459          29          0.8        1242.6       0.1X
+timestamp strings                                  1508           1519           9          6.6         150.8       0.7X
+parse timestamps from Dataset[String]             21674          21997         455          0.5        2167.4       0.1X
+infer timestamps from Dataset[String]             42141          42230         105          0.2        4214.1       0.0X
+date strings                                       1694           1701           8          5.9         169.4       0.7X
+parse dates from Dataset[String]                  12929          12951          25          0.8        1292.9       0.1X
+from_csv(timestamp)                               20603          20786         166          0.5        2060.3       0.1X
+from_csv(date)                                    12325          12338          12          0.8        1232.5       0.1X
 
-Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.2
-Intel(R) Core(TM) i7-4850HQ CPU @ 2.30GHz
+Java HotSpot(TM) 64-Bit Server VM 1.8.0_231-b11 on Mac OS X 10.15.4
+Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
 Filters pushdown:                         Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------
-w/o filters                                       12557          12634          78          0.0      125572.9       1.0X
-pushdown disabled                                 12449          12509          65          0.0      124486.4       1.0X
-w/ filters                                         1372           1393          18          0.1       13724.8       9.1X
+w/o filters                                       12455          12474          22          0.0      124553.8       1.0X
+pushdown disabled                                 12462          12486          29          0.0      124624.9       1.0X
+w/ filters                                         1073           1092          18          0.1       10727.6      11.6X